+++ modules/apps/src/WithdrawApp/index.ts
@@ -1,3 +1,4 @@
 export * from "./convert";
 export * from "./registry";
 export * from "./validation";
+export * from "./withdrawCommitment";
