+++ modules/cf-core/src/ethereum/index.ts
@@ -8,5 +8,4 @@ export {
 } from "./free-balance-app";
 export { SetStateCommitment } from "./set-state-commitment";
 export { SetupCommitment } from "./setup-commitment";
-export { WithdrawERC20Commitment } from "./withdraw-erc20-commitment";
-export { WithdrawETHCommitment } from "./withdraw-eth-commitment";
+export { MultisigCommitment } from "./multisig-commitment";
