+++ modules/cf-core/src/ethereum/multisig-commitment.ts
@@ -5,10 +5,8 @@ import { CFCoreTypes, EthereumCommitment, MultisigTransaction } from "../types";
 import { sortSignaturesBySignerAddress, sortStringSignaturesBySignerAddress } from "../utils";
 
 /// A commitment to make MinimumViableMultisig perform a message call
-export abstract class MultisigCommitment extends EthereumCommitment {
-  constructor(readonly multisigAddress: string, readonly multisigOwners: string[]) {
-    super();
-  }
+export abstract class MultisigCommitment implements EthereumCommitment {
+  constructor(readonly multisigAddress: string, readonly multisigOwners: string[]) {}
 
   abstract getTransactionDetails(): MultisigTransaction;
 
@@ -16,14 +14,12 @@ export abstract class MultisigCommitment extends EthereumCommitment {
     const multisigInput = this.getTransactionDetails();
     let signaturesList: string[];
 
-    if(typeof sigs[0] == "string") {
+    if (typeof sigs[0] == "string") {
       //@ts-ignore
       signaturesList = sortStringSignaturesBySignerAddress(this.hashToSign(), sigs);
     } else {
       //@ts-ignore
-      signaturesList = sortSignaturesBySignerAddress(this.hashToSign(), sigs).map(
-        joinSignature,
-      );
+      signaturesList = sortSignaturesBySignerAddress(this.hashToSign(), sigs).map(joinSignature);
     }
 
     const txData = new Interface(MinimumViableMultisig.abi).functions.execTransaction.encode([
