+++ modules/cf-core/src/types.ts
@@ -1,44 +1,8 @@
-import {
-  CFCoreTypes,
-  ILoggerService,
-  NetworkContext,
-  ProtocolMessage,
-  SolidityValueType,
-} from "@connext/types";
+import { ILoggerService, NetworkContext, ProtocolMessage, SolidityValueType } from "@connext/types";
 import { BaseProvider } from "ethers/providers";
-import { Signature } from "ethers/utils";
 
 import { Opcode } from "./machine";
 import { StateChannel } from "./models";
-
-export abstract class EthereumCommitment {
-  // todo(xuanji): EthereumCommitment was designed under the assumption that
-  // `hashToSign` returns the same hash for different signers. However, in the
-  // install-virtual-app protocol, the hash that the intermediary signs is
-  // different from the one the other participants sign. The optional
-  // `signerIsIntermediary` flag is a hack that is only used by the
-  // `install-virtual-app protocol`. `intermediarySignature` in `transaction`
-  // is the same kind of hack.
-  public abstract hashToSign(signerIsIntermediary?: boolean): string;
-  public abstract getSignedTransaction(
-    signatures: Signature[],
-    intermediarySignature?: Signature,
-  ): CFCoreTypes.MinimalTransaction;
-}
-
-export enum MultisigOperation {
-  Call = 0,
-  DelegateCall = 1,
-  // Gnosis Safe uses "2" for CREATE, but we don't actually
-  // make use of it in our code. Still, I put this here to be
-  // maximally explicit that we based the data structure on
-  // Gnosis's implementation of a Multisig
-  Create = 2,
-}
-
-export type MultisigTransaction = CFCoreTypes.MinimalTransaction & {
-  operation: MultisigOperation;
-};
 
 export type ProtocolExecutionFlow = {
   [x: number]: (context: Context) => AsyncIterableIterator<any[]>;
@@ -117,4 +81,7 @@ export {
   UpdateProtocolParams,
   UpdateStateMessage,
   virtualAppAgreementEncoding,
+  EthereumCommitment,
+  MultisigOperation,
+  MultisigTransaction,
 } from "@connext/types";
